refactor(inspection-service): extract inspection-status sub-service
Move the inspection + report status-machine transitions (confirm / cancel /
uncancel / submitReport / returnReport / unpublishReport / markPaymentReceived)
plus the private fetchForStatusChange helper into InspectionStatusService. Self
contained (no cross-service calls); the facade delegates. Moved code is
byte-identical.
